requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiokafka==0.9.0
aiosignal==1.3.1
async-timeout==4.0.3
asynch==0.2.3
asyncpg==0.29.0
attrs==23.2.0
backoff==2.2.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
croniter==2.0.5
cryptography==42.0.7
dask==2024.5.2
faust-streaming==0.11.0
frozenlist==1.4.1
fsspec==2024.5.0
h5py==3.11.0
idna==3.7
ijson==3.2.3
importlib_metadata==7.1.0
intervaltree==3.1.0
kafka-python==1.4.6
leb128==1.0.7
locket==1.0.0
lz4==4.3.3
mode-streaming==0.4.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
opentracing==2.4.0
packaging==24.0
pandas==2.2.2
partd==1.4.2
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-rapidjson==1.17
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sortedcontainers==2.4.0
terminaltables==3.1.10
toolz==0.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
venusian==3.1.0
websockets==12.0
yarl==1.9.4
zipp==3.19.1
zstd==1.5.5.1
